update PR feedback
