+++ verify-cborld-presentation/.eslintrc
@@ -1,7 +1,7 @@
 {
   "root": true,
   "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint", "import", "functional", "eslint-plugin-tsdoc"],
+  "plugins": ["@typescript-eslint", "import", "functional", "eslint-plugin-tsdoc", "react-hooks"],
   "extends": [
     "eslint:recommended",
     "plugin:@typescript-eslint/eslint-recommended",
@@ -10,6 +10,8 @@
   ],
   "ignorePatterns": ["node_modules", "lib", "dist", "build"],
   "rules": {
+    "react-hooks/rules-of-hooks": "error", // Checks rules of Hooks
+    "react-hooks/exhaustive-deps": "error", // Checks effect dependencies
     "tsdoc/syntax": "warn",
     "@typescript-eslint/explicit-function-return-type": [
       "error",
