+++ verify-cborld-presentation/package.json
@@ -9,12 +9,14 @@
     "@types/react": "17.0.0",
     "@types/react-dom": "17.0.0",
     "@types/react-router-dom": "5.1.7",
+    "@types/styled-components": "5.1.10",
     "@typescript-eslint/eslint-plugin": "4.20.0",
     "@typescript-eslint/parser": "4.20.0",
     "base32-decode": "1.0.0",
     "eslint": "7.23.0",
     "eslint-plugin-functional": "3.2.1",
     "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-react-hooks": "4.2.0",
     "eslint-plugin-tsdoc": "0.2.14",
     "http-proxy-middleware": "2.0.0",
     "husky": "3.0.4",
