+++ verify-cborld-presentation/src/pages/Qr.tsx
