+++ verify-cborld-presentation/src/pages/index.ts
@@ -1,4 +1,4 @@
-export { HomePage } from "./home";
-export { ScanPage } from "./scan";
-export { ResultPage } from "./result";
-export { QrPage } from "./qr";
+export { HomePage } from "./Home";
+export { ScanPage } from "./Scan";
+export { ResultPage } from "./Result";
+export { QrPage } from "./Qr";
